pyproject.toml
[project]
name = "wedding-ai-pipeline"
version = "0.1.0"
description = "Computational Architectures for Automated Wedding Photography Post-Production"
authors = [
    {name = "Wedding AI Team", email = "dev@example.com"}
]
dependencies = [
    "rawpy>=0.19",
    "opencv-python-headless>=4.8",
    "numpy>=1.24",
    "Pillow>=10.0",
    "imagehash>=4.3",
    "mediapipe>=0.10",
    "insightface>=0.7",
    "gfpgan>=1.3",
    "torch>=2.1",
    "torchvision>=0.16",
    "transformers>=4.36",
    "open-clip-torch>=2.24",
    "scikit-learn>=1.3",
    "reportlab>=4.0",
    "pyyaml>=6.0",
    "pydantic>=2.5",
    "click>=8.1",
    "rich>=13.0",
    "llama-cpp-python>=0.2",
    "pybind11>=2.11",
    "pytest>=7.4",
]
readme = "README.md"
requires-python = ">=3.11"

[build-system]
requires = ["setuptools>=61.0", "pybind11>=2.11"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
